+++ .gitignore
@@ -2,3 +2,4 @@
 */__pycache__/*
 __pycache__/*
 .gitignore
+HW_6_Files/temporary_script.py
